requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
pypng==0.20220715.0
pyzbar==0.1.9
qrcode==7.4.2
typing_extensions==4.12.2
Werkzeug==3.0.4
